+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 # Change log
 
-## [v1.1.3](https://github.com/simvue-io/client/releases/tag/v1.1.3) - 2024-12-09
+## Unreleased
 
 * Fixed bug with `requirements.txt` metadata read.
 * Added Simvue server version check.
+* Remove checking of server version in offline mode and add default run mode to configuration options.
 
 ## [v1.1.2](https://github.com/simvue-io/client/releases/tag/v1.1.2) - 2024-11-06