+++ CHANGELOG.md
@@ -1,10 +1,16 @@
 # Change log
 
 ## Unreleased
+
 * Fixed bug with `requirements.txt` metadata read.
+* Added Simvue server version check.
+* Remove checking of server version in offline mode and add default run mode to configuration options.
+* Fix offline mode class initialisation, and propagation of configuration.
+
 ## [v1.1.2](https://github.com/simvue-io/client/releases/tag/v1.1.2) - 2024-11-06
 
 * Fix bug in offline mode directory retrieval.
+
 ## [v1.1.1](https://github.com/simvue-io/client/releases/tag/v1.1.1) - 2024-10-22
 
 * Add missing `offline.cache` key to TOML config.